+++ server/routes/sitemap.xml.ts
@@ -1,0 +1,34 @@
+const siteUrl = 'https://askara.basim.id'
+const lastModified = '2026-07-07'
+
+const routes = [{ loc: '/', priority: '1.0', changefreq: 'weekly' }]
+
+function escapeXml(value: string) {
+  return value
+    .replaceAll('&', '&amp;')
+    .replaceAll('<', '&lt;')
+    .replaceAll('>', '&gt;')
+    .replaceAll('"', '&quot;')
+    .replaceAll("'", '&apos;')
+}
+
+export default defineEventHandler((event) => {
+  setHeader(event, 'content-type', 'application/xml; charset=utf-8')
+
+  const urls = routes
+    .map(
+      (route) => `  <url>
+    <loc>${escapeXml(`${siteUrl}${route.loc}`)}</loc>
+    <lastmod>${lastModified}</lastmod>
+    <changefreq>${route.changefreq}</changefreq>
+    <priority>${route.priority}</priority>
+  </url>`,
+    )
+    .join('\n')
+
+  return `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+${urls}
+</urlset>
+`
+})